old_code/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==2.0.0
art==5.9
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
blinker==1.6.2
Brlapi==0.8.4
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Columnar==1.4.1
command-not-found==0.3
contourpy==1.0.7
cryptography==38.0.4
cs50==9.2.5
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
defer==1.0.6
distlib==0.3.6
distro==1.8.0
distro-info==1.5
docutils==0.19
docx==0.2.4
docxcompose==1.4.0
docxtpl==0.16.7
duplicity==0.8.22
exceptiongroup==1.1.1
fasteners==0.17.3
filelock==3.12.0
flake8==5.0.4
Flask==2.3.2
fonttools==4.39.3
fpdf==1.7.2
frozenlist==1.3.3
future==0.18.2
ghp-import==2.1.0
google-api-python-client==1.7.12
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==2.0.1
h11==0.14.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
keyring==23.9.3
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
livereload==2.6.3
lockfile==0.12.2
louis==3.24.0
lunr==0.6.2
lxml==4.9.2
Mako==1.2.4.dev0
Markdown==3.4.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplot==0.1.9
matplotlib==3.7.1
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.4.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
nltk==3.8
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
openai==0.27.4
outcome==1.2.0
packaging==23.1
pandas==2.0.1
paramiko==2.12.0
pathspec==0.11.1
pexpect==4.8.0
Pillow==9.5.0
pipx==1.1.0
pkginfo==1.9.6
platformdirs==3.5.1
pluggy==1.0.0+repack
prettytable==3.7.0
psutil==5.9.4
psycopg-binary==3.1.9
psycopg2==2.9.6
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.10.0
pycups==2.0.1
PyDrive2==0.0.0
pyflakes==2.5.0
Pygments==2.14.0
PyGObject==3.44.1
pyinotify==0.9.6
PyJWT==2.6.0
pyloco==0.0.139
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pypandoc==1.11
pyparsing==3.0.9
PyPDF2==3.0.1
PySimpleGUI==4.60.5
PySocks==1.7.1
pytest==7.3.1
python-apt==2.5.3+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2022.7.1
pyxdg==0.28
PyYAML==6.0
pyyaml_env_tag==0.1
readme-renderer==37.3
regex==2022.10.31
reportlab==3.6.12
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.1
rsa==4.8
SecretStorage==3.3.3
selenium==4.10.0
simplejson==3.18.3
SimpleWebSocketServer==0.1.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.46
sqlparse==0.4.4
string-color==1.2.3
systemd-python==235
termcolor==2.3.0
termcolor2==0.0.3
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.10.3
twine==4.0.2
typing==3.7.4.3
typing_extensions==4.6.2
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
unoconv==0.9.0
uritemplate==4.1.1
urllib3==1.26.12
usb-creator==0.3.16
userpath==1.8.0
ushlex==0.99.1
virtualenv==20.23.0
wadllib==1.3.6
watchdog==2.2.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.3
wsproto==1.2.0
wxPython==4.2.0
xdg==5
xkit==0.0.0
yapf==0.33.0
yarl==1.9.1
zipp==1.0.0
